revert frida-gum bump. Current frida-gum doesn't build on iOS :/

+++ libafl_frida/src/helper.rs
@@ -304,7 +304,7 @@ where
                                     basereg,
                                     indexreg,
                                     scale,
-                                    disp as isize,
+                                    disp,
                                 );
                             }
                         }
